Fix: fix the 11th move freezing bug
The reason that the game froze after 10 moves was because the function
that oversaw the validity of computer's move (isValidShot function)
checked to see if the new shot had a unique x AND y value not Or.
Which meant after 10 moves there could be no shot with a unique
x and y because all unique x and y values would be taken. What happened
next was that the computerAttack function kept stuck passing in
different values infinitely and this caused the page to freeze.